+++ src/mrinfo.c
@@ -86,6 +86,7 @@
 
 uint32_t  our_addr, target_addr = 0;	/* in NET order */
 int       debug = 0;
+int       mrt_table_id = 0;		/* dummy, unused */
 int       nflag = 0;
 int       retries = DEFAULT_RETRIES;
 int       timeout = DEFAULT_TIMEOUT;