+++ ChangeLog.md
@@ -10,6 +10,7 @@ All notable changes to the project are documented in this file.
 - Add support for `-i,--ident=NAME` to change identity of an instance
 - Add support for `-p,--pidfile=FILE` to override default PID file
 - Touch PID file at SIGHUP to acknowledge done reloading .conf file
+- Add support for `-t,--table-id=ID`, multicast routing tables (Linux)
 - Add support for `-u,--ipc=FILE` to override `/var/run/mrouted.sock`
   file, used for communication with `mroutectl`